% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aargh-package.R, R/aargh.R
\docType{package}
\name{aargh}
\alias{aargh}
\alias{aargh-package}
\alias{aargh}
\title{aargh.}
\usage{
aargh(f)
}
\arguments{
\item{f}{the function to wrap}
}
\description{
aargh.

Wraps a function for command line use
}
