% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aargh.R
\name{argparse_data_from_formal}
\alias{argparse_data_from_formal}
\title{Interrogates a formal to get argparse data}
\usage{
argparse_data_from_formal(x)
}
\arguments{
\item{x}{a formal of a function}
}
\description{
Interrogates a formal to get argparse data
}
